LayoutTests/http/tests/site-isolation/wheel-event-coordinates-expected.txt

Verifies that the iframe receives wheel events with the expected coordinates.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.data is '40,54'
PASS successfullyParsed is true

TEST COMPLETE
